% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_fcts_pdf.R
\name{add_tex_content}
\alias{add_tex_content}
\title{Replaces content in tex template file}
\usage{
add_tex_content(
  f_in,
  f_out,
  student_name,
  student_version,
  student_id,
  n_q,
  class_name,
  exercise_name,
  exam_links
)
}
\arguments{
\item{f_in}{File with tex code}

\item{student_name}{Name of student}

\item{student_version}{Version of student}

\item{n_q}{Number of questions in exercise}

\item{class_name}{Name of class}

\item{exercise_name}{Name of exercise}

\item{exam_links}{Links to add}

\item{dir_out}{Folder out}
}
\value{
TRUE, if sucessfull
}
\description{
Used for replacing names, id, version and so on on tex exercise files.
}
\examples{
\dontrun{
afedR_add_html_content(f_in = 'example.html', dir_out = tempdir(),
                       student_name = 'George', student_version = 1,
                       n_q = 10, class_name = 'example class', exercise_name = 'sample',
                       exam_links = NA)
}
}
