% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_fcts.R
\name{afedR_calc_ret}
\alias{afedR_calc_ret}
\title{Calculates a return vector}
\usage{
afedR_calc_ret(P, tickers = rep("ticker", length(P)))
}
\arguments{
\item{P}{Vector of prices}

\item{tickers}{Vector of tickers}
}
\value{
A vector of returns
}
\description{
Calculates a return vector
}
\examples{
print(afedR_calc_ret(1:10))
}
