% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R
\name{afedR_get_book_files}
\alias{afedR_get_book_files}
\title{Copy all book files to local folder}
\usage{
afedR_get_book_files(path_to_copy = "~")
}
\arguments{
\item{path_to_copy}{Path to copy all the book files}
}
\value{
TRUE, if sucessful
}
\description{
This function will grab files from the afedR package and copy all of it to a local folder,
separated by directories including R code, data, slides and end-chapter exercises.
}
\examples{
\dontrun{
flag <- afedR_get_book_files()
}
}
