% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation_fcts.R
\name{afedR_replace_outliers}
\alias{afedR_replace_outliers}
\title{Replaces outliers from a vector and returns a new vector}
\usage{
afedR_replace_outliers(col_in, my_prob = 0.05)
}
\arguments{
\item{col_in}{The vector}

\item{my_prob}{Probability of quantiles (will remove quantiles at p and 1-p)}
}
\value{
A vector without the outliers
}
\description{
Replaces outliers from a vector and returns a new vector
}
\examples{
x <- afedR_replace_outliers(runif(100))
}
