% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_fcts.R
\name{build_answer_key}
\alias{build_answer_key}
\title{Builds answer key}
\usage{
build_answer_key(my_exam, students_names)
}
\arguments{
\item{my_exam}{A single exam list}

\item{students_names}{Names of students}
}
\value{
A dataframe
}
\description{
Builds answer key
}
