% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_fcts.R
\name{build_answers_text}
\alias{build_answers_text}
\title{Creates text for random questions}
\usage{
build_answers_text(text1, text2, text3)
}
\arguments{
\item{text1}{string 1}

\item{text2}{string 2}

\item{text3}{string 3}
}
\value{
A list
}
\description{
Creates text for random questions
}
