% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_fcts_pdf.R
\name{build_pdf_exercises}
\alias{build_pdf_exercises}
\title{Builds pdf exercices}
\usage{
build_pdf_exercises(
  students_names,
  students_ids,
  class_name,
  exercise_name,
  exercises_to_compile,
  chapters_to_include,
  dir_out = "adfeR-pdf-exercises",
  solution = FALSE
)
}
\arguments{
\item{students_names}{Names of students (a vector)}

\item{students_ids}{Ids of students (a vector)}

\item{class_name}{The name of the class}

\item{exercise_name}{The name of the exercises}

\item{chapters_to_include}{Chapter to include in exercise (1-13)}

\item{dir_out}{Folder to copy exercise html files}
}
\value{
a list
}
\description{
Builds pdf exercices
}
