% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_webex_fcts.R
\name{check_answers}
\alias{check_answers}
\title{Check exams answers}
\usage{
check_answers(answers_in)
}
\arguments{
\item{answers_in}{Char vector}
}
\value{
TRUE or FALSE
}
\description{
Check exams answers
}
