% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_formats.R
\name{format_cash}
\alias{format_cash}
\title{formats cash}
\usage{
format_cash(x, type_cash = "USD")
}
\arguments{
\item{x}{numeric}

\item{type_cash}{BRL or USD}
}
\value{
cash string
}
\description{
formats cash
}
