% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_fcts.R, R/exams_webex_fcts.R
\name{make_random_answers}
\alias{make_random_answers}
\title{Title}
\usage{
make_random_answers(solution, candidates = NA, is_cash = FALSE)

make_random_answers(solution, candidates = NA, is_cash = FALSE)
}
\arguments{
\item{solution}{The solution}

\item{candidates}{Candidates (opcional)}

\item{is_cash}{Where solution is cash}
}
\value{


Char vector with randomized answers
}
\description{
Title

Creates random answers
}
